Refactor few widgets to dedicated modules
ResizingPanel, Pages, EditField, HotkeyLabel, Label and Scrollbar
