+++ library/lua/gui/widgets.lua
@@ -11,17 +11,18 @@ Widget = require('gui.widgets.widget')
 Divider = require('gui.widgets.divider')
 Panel = require('gui.widgets.panel')
 Window = require('gui.widgets.window')
+ResizingPanel = require('gui.widgets.resizing_panel')
+Pages = require('gui.widgets.pages')
+EditField = require('gui.widgets.edit_field')
+HotkeyLabel = require('gui.widgets.hotkey_label')
+Label = require('gui.widgets.label')
+Scrollbar = require('gui.widgets.scrollbar')
+
+makeButtonLabelText = Label.makeButtonLabelText
+parse_label_text = Label.parse_label_text
 
 local getval = utils.getval
 local to_pen = dfhack.pen.parse
-
----@param view gui.View
----@param vis boolean
-local function show_view(view,vis)
-    if view then
-        view.visible = vis
-    end
-end
 
 ---@param tab? table
 ---@param idx integer
@@ -32,1132 +33,6 @@ local function map_opttab(tab,idx)
     else
         return idx
     end
-end
-
----@enum STANDARDSCROLL
-STANDARDSCROLL = {
-    STANDARDSCROLL_UP = -1,
-    KEYBOARD_CURSOR_UP = -1,
-    STANDARDSCROLL_DOWN = 1,
-    KEYBOARD_CURSOR_DOWN = 1,
-    STANDARDSCROLL_PAGEUP = '-page',
-    KEYBOARD_CURSOR_UP_FAST = '-page',
-    STANDARDSCROLL_PAGEDOWN = '+page',
-    KEYBOARD_CURSOR_DOWN_FAST = '+page',
-}
-
--------------------
--- ResizingPanel --
--------------------
-
----@class widgets.ResizingPanel.attrs: widgets.Panel.attrs
----@field auto_height boolean
----@field auto_width boolean
-
----@class widgets.ResizingPanel.attrs.partial: widgets.ResizingPanel.attrs
-
----@class widgets.ResizingPanel: widgets.Panel, widgets.ResizingPanel.attrs
----@field super widgets.Panel
----@field ATTRS widgets.ResizingPanel.attrs|fun(attributes: widgets.ResizingPanel.attrs.partial)
----@overload fun(init_table: widgets.ResizingPanel.attrs.partial): self
-ResizingPanel = defclass(ResizingPanel, Panel)
-
-ResizingPanel.ATTRS{
-    auto_height = true,
-    auto_width = false,
-}
-
--- adjust our frame dimensions according to positions and sizes of our subviews
-function ResizingPanel:postUpdateLayout(frame_body)
-    local w, h = 0, 0
-    for _,s in ipairs(self.subviews) do
-        if getval(s.visible) then
-            w = math.max(w, (s.frame and s.frame.l or 0) +
-                            (s.frame and s.frame.w or frame_body.width))
-            h = math.max(h, (s.frame and s.frame.t or 0) +
-                            (s.frame and s.frame.h or frame_body.height))
-        end
-    end
-    local l,t,r,b = gui.parse_inset(self.frame_inset)
-    w = w + l + r
-    h = h + t + b
-    if self.frame_style then
-        w = w + 2
-        h = h + 2
-    end
-    if not self.frame then self.frame = {} end
-    local oldw, oldh = self.frame.w, self.frame.h
-    if not self.auto_height then h = oldh end
-    if not self.auto_width then w = oldw end
-    self.frame.w, self.frame.h = w, h
-    if not self._updateLayoutGuard and (oldw ~= w or oldh ~= h) then
-        self._updateLayoutGuard = true -- protect against infinite loops
-        self:updateLayout() -- our frame has changed, we need to fully refresh
-    end
-    self._updateLayoutGuard = nil
-end
-
------------
--- Pages --
------------
-
----@class widgets.Pages.initTable: widgets.Panel.attrs
----@field selected? integer|string
-
----@class widgets.Pages: widgets.Panel
----@field super widgets.Panel
----@overload fun(attributes: widgets.Pages.initTable): self
-Pages = defclass(Pages, Panel)
-
----@param self widgets.Pages
----@param args widgets.Pages.initTable
-function Pages:init(args)
-    for _,v in ipairs(self.subviews) do
-        v.visible = false
-    end
-    self:setSelected(args.selected or 1)
-end
-
----@param idx integer|string
-function Pages:setSelected(idx)
-    if type(idx) ~= 'number' then
-        local key = idx
-        if type(idx) == 'string' then
-            key = self.subviews[key]
-        end
-        idx = utils.linear_index(self.subviews, key)
-        if not idx then
-            error('Unknown page: '..tostring(key))
-        end
-    end
-
-    show_view(self.subviews[self.selected], false)
-    self.selected = math.min(math.max(1, idx), #self.subviews)
-    show_view(self.subviews[self.selected], true)
-end
-
----@return integer index
----@return gui.View child
-function Pages:getSelected()
-    return self.selected, self.subviews[self.selected]
-end
-
----@return gui.View child
-function Pages:getSelectedPage()
-    return self.subviews[self.selected]
-end
-
-----------------
--- Edit field --
-----------------
-
----@class widgets.EditField.attrs: widgets.Widget.attrs
----@field label_text? string
----@field text string
----@field text_pen? dfhack.color|dfhack.pen
----@field on_char? function
----@field on_change? function
----@field on_submit? function
----@field on_submit2? function
----@field key? string
----@field key_sep? string
----@field modal boolean
----@field ignore_keys? string[]
-
----@class widgets.EditField.attrs.partial: widgets.EditField.attrs
-
----@class widgets.EditField: widgets.Widget, widgets.EditField.attrs
----@field super widgets.Widget
----@field ATTRS widgets.EditField.attrs|fun(attributes: widgets.EditField.attrs.partial)
----@overload fun(init_table: widgets.EditField.attrs.partial): self
-EditField = defclass(EditField, Widget)
-
-EditField.ATTRS{
-    label_text = DEFAULT_NIL,
-    text = '',
-    text_pen = DEFAULT_NIL,
-    on_char = DEFAULT_NIL,
-    on_change = DEFAULT_NIL,
-    on_submit = DEFAULT_NIL,
-    on_submit2 = DEFAULT_NIL,
-    key = DEFAULT_NIL,
-    key_sep = DEFAULT_NIL,
-    modal = false,
-    ignore_keys = DEFAULT_NIL,
-}
-
----@param init_table widgets.EditField.attrs
-function EditField:preinit(init_table)
-    init_table.frame = init_table.frame or {}
-    init_table.frame.h = init_table.frame.h or 1
-end
-
-function EditField:init()
-    local function on_activate()
-        self.saved_text = self.text
-        self:setFocus(true)
-    end
-
-    self.start_pos = 1
-    self.cursor = #self.text + 1
-
-    self:addviews{HotkeyLabel{frame={t=0,l=0},
-                              key=self.key,
-                              key_sep=self.key_sep,
-                              label=self.label_text,
-                              on_activate=self.key and on_activate or nil}}
-end
-
-function EditField:getPreferredFocusState()
-    return not self.key
-end
-
-function EditField:setCursor(cursor)
-    if not cursor or cursor > #self.text then
-        self.cursor = #self.text + 1
-        return
-    end
-    self.cursor = math.max(1, cursor)
-end
-
-function EditField:setText(text, cursor)
-    local old = self.text
-    self.text = text
-    self:setCursor(cursor)
-    if self.on_change and text ~= old then
-        self.on_change(self.text, old)
-    end
-end
-
-function EditField:postUpdateLayout()
-    self.text_offset = self.subviews[1]:getTextWidth()
-end
-
----@param dc gui.Painter
-function EditField:onRenderBody(dc)
-    dc:pen(self.text_pen or COLOR_LIGHTCYAN)
-
-    local cursor_char = '_'
-    if not getval(self.active) or not self.focus or gui.blink_visible(300) then
-        cursor_char = (self.cursor > #self.text) and ' ' or
-                                        self.text:sub(self.cursor, self.cursor)
-    end
-    local txt = self.text:sub(1, self.cursor - 1) .. cursor_char ..
-                                                self.text:sub(self.cursor + 1)
-    local max_width = dc.width - self.text_offset
-    self.start_pos = 1
-    if #txt > max_width then
-        -- get the substring in the vicinity of the cursor
-        max_width = max_width - 2
-        local half_width = math.floor(max_width/2)
-        local start_pos = math.max(1, self.cursor-half_width)
-        local end_pos = math.min(#txt, self.cursor+half_width-1)
-        if self.cursor + half_width > #txt then
-            start_pos = #txt - (max_width - 1)
-        end
-        if self.cursor - half_width <= 1 then
-            end_pos = max_width + 1
-        end
-        self.start_pos = start_pos > 1 and start_pos - 1 or start_pos
-        txt = ('%s%s%s'):format(start_pos == 1 and '' or string.char(27),
-                                txt:sub(start_pos, end_pos),
-                                end_pos == #txt and '' or string.char(26))
-    end
-    dc:advance(self.text_offset):string(txt)
-end
-
-function EditField:insert(text)
-    local old = self.text
-    self:setText(old:sub(1,self.cursor-1)..text..old:sub(self.cursor),
-                 self.cursor + #text)
-end
-
-function EditField:onInput(keys)
-    if not self.focus then
-        -- only react to our hotkey
-        return self:inputToSubviews(keys)
-    end
-
-    if self.ignore_keys then
-        for _,ignore_key in ipairs(self.ignore_keys) do
-            if keys[ignore_key] then return false end
-        end
-    end
-
-    if self.key and (keys.LEAVESCREEN or keys._MOUSE_R) then
-        self:setText(self.saved_text)
-        self:setFocus(false)
-        return true
-    end
-
-    if keys.SELECT or keys.SELECT_ALL then
-        if self.key then
-            self:setFocus(false)
-        end
-        if keys.SELECT_ALL then
-            if self.on_submit2 then
-                self.on_submit2(self.text)
-                return true
-            end
-        else
-            if self.on_submit then
-                self.on_submit(self.text)
-                return true
-            end
-        end
-        return not not self.key
-    elseif keys._STRING then
-        local old = self.text
-        if keys._STRING == 0 then
-            -- handle backspace
-            local del_pos = self.cursor - 1
-            if del_pos > 0 then
-                self:setText(old:sub(1, del_pos-1) .. old:sub(del_pos+1),
-                             del_pos)
-            end
-        else
-            local cv = string.char(keys._STRING)
-            if not self.on_char or self.on_char(cv, old) then
-                self:insert(cv)
-            elseif self.on_char then
-                return self.modal
-            end
-        end
-        return true
-    elseif keys.KEYBOARD_CURSOR_LEFT then
-        self:setCursor(self.cursor - 1)
-        return true
-    elseif keys.CUSTOM_CTRL_B then -- back one word
-        local _, prev_word_end = self.text:sub(1, self.cursor-1):
-                                               find('.*[%w_%-][^%w_%-]')
-        self:setCursor(prev_word_end or 1)
-        return true
-    -- commented out until we get HOME key support from DF
-    -- elseif keys.CUSTOM_CTRL_A then -- home
-    --     self:setCursor(1)
-    --     return true
-    elseif keys.KEYBOARD_CURSOR_RIGHT then
-        self:setCursor(self.cursor + 1)
-        return true
-    elseif keys.CUSTOM_CTRL_F then -- forward one word
-        local _,next_word_start = self.text:find('[^%w_%-][%w_%-]', self.cursor)
-        self:setCursor(next_word_start)
-        return true
-    elseif keys.CUSTOM_CTRL_E then -- end
-        self:setCursor()
-        return true
-    elseif keys.CUSTOM_CTRL_C then
-        dfhack.internal.setClipboardTextCp437(self.text)
-        return true
-    elseif keys.CUSTOM_CTRL_X then
-        dfhack.internal.setClipboardTextCp437(self.text)
-        self:setText('')
-        return true
-    elseif keys.CUSTOM_CTRL_V then
-        self:insert(dfhack.internal.getClipboardTextCp437())
-        return true
-    elseif keys._MOUSE_L_DOWN then
-        local mouse_x = self:getMousePos()
-        if mouse_x then
-            self:setCursor(self.start_pos + mouse_x - (self.text_offset or 0))
-            return true
-        end
-    end
-
-    -- if we're modal, then unconditionally eat all the input
-    return self.modal
-end
-
----------------
--- Scrollbar --
----------------
-
-SCROLL_INITIAL_DELAY_MS = 300
-SCROLL_DELAY_MS = 20
-
----@class widgets.Scrollbar.attrs: widgets.Widget.attrs
----@field on_scroll? fun(new_top_elem?: integer)
-
----@class widgets.Scrollbar.attrs.partial: widgets.Scrollbar.attrs
-
----@class widgets.Scrollbar: widgets.Widget, widgets.Scrollbar.attrs
----@field super widgets.Widget
----@field ATTRS widgets.Scrollbar.attrs|fun(attributes: widgets.Scrollbar.attrs.partial)
----@overload fun(init_table: widgets.Scrollbar.attrs.partial): self
-Scrollbar = defclass(Scrollbar, Widget)
-
-Scrollbar.ATTRS{
-    on_scroll = DEFAULT_NIL,
-}
-
----@param init_table widgets.Scrollbar.attrs.partial
-function Scrollbar:preinit(init_table)
-    init_table.frame = init_table.frame or {}
-    init_table.frame.w = init_table.frame.w or 2
-end
-
-function Scrollbar:init()
-    self.last_scroll_ms = 0
-    self.is_first_click = false
-    self.scroll_spec = nil
-    self.is_dragging = false -- index of the scrollbar tile that we're dragging
-    self:update(1, 1, 1)
-end
-
-local function scrollbar_get_max_pos_and_height(scrollbar)
-    local frame_body = scrollbar.frame_body
-    local scrollbar_body_height = (frame_body and frame_body.height or 3) - 2
-
-    local height = math.max(2, math.floor(
-        (math.min(scrollbar.elems_per_page, scrollbar.num_elems) * scrollbar_body_height) /
-        scrollbar.num_elems))
-
-    return scrollbar_body_height - height, height
-end
-
--- calculate and cache the number of tiles of empty space above the top of the
--- scrollbar and the number of tiles the scrollbar should occupy to represent
--- the percentage of text that is on the screen.
--- if elems_per_page or num_elems are not specified, the last values passed to
--- Scrollbar:update() are used.
-function Scrollbar:update(top_elem, elems_per_page, num_elems)
-    if not top_elem then error('must specify index of new top element') end
-    elems_per_page = elems_per_page or self.elems_per_page
-    num_elems = num_elems or self.num_elems
-    self.top_elem = top_elem
-    self.elems_per_page, self.num_elems = elems_per_page, num_elems
-
-    local max_pos, height = scrollbar_get_max_pos_and_height(self)
-    local pos = (num_elems == elems_per_page) and 0 or
-            math.ceil(((top_elem-1) * max_pos) /
-                      (num_elems - elems_per_page))
-
-    self.bar_offset, self.bar_height = pos, height
-end
-
-local function scrollbar_do_drag(scrollbar)
-    local x,y = dfhack.screen.getMousePos()
-    if not y then return end
-    x,y = scrollbar.frame_body:localXY(x, y)
-    local cur_pos = y - scrollbar.is_dragging
-    local max_top = scrollbar.num_elems - scrollbar.elems_per_page + 1
-    local max_pos = scrollbar_get_max_pos_and_height(scrollbar)
-    local new_top_elem = math.floor(cur_pos * max_top / max_pos) + 1
-    new_top_elem = math.max(1, math.min(new_top_elem, max_top))
-    if new_top_elem ~= scrollbar.top_elem then
-        scrollbar.on_scroll(new_top_elem)
-    end
-end
-
-local function scrollbar_is_visible(scrollbar)
-    return scrollbar.elems_per_page < scrollbar.num_elems
-end
-
-local SBSO = df.global.init.scrollbar_texpos[0] --Scroll Bar Spritesheet Offset / change this to point to a different spritesheet (ui themes, anyone? :p)
-local SCROLLBAR_UP_LEFT_PEN = to_pen{tile=SBSO+0, ch=47, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_UP_RIGHT_PEN = to_pen{tile=SBSO+1, ch=92, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_DOWN_LEFT_PEN = to_pen{tile=SBSO+24, ch=92, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_DOWN_RIGHT_PEN = to_pen{tile=SBSO+25, ch=47, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_UP_LEFT_PEN = to_pen{tile=SBSO+6, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_UP_RIGHT_PEN = to_pen{tile=SBSO+7, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_LEFT_PEN = to_pen{tile=SBSO+30, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_RIGHT_PEN = to_pen{tile=SBSO+31, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_UP_LEFT_PEN = to_pen{tile=SBSO+10, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_UP_RIGHT_PEN = to_pen{tile=SBSO+11, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_DOWN_LEFT_PEN = to_pen{tile=SBSO+22, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_DOWN_RIGHT_PEN = to_pen{tile=SBSO+23, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_LEFT_PEN = to_pen{tile=SBSO+18, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_RIGHT_PEN = to_pen{tile=SBSO+19, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_DOWN_LEFT_PEN = to_pen{tile=SBSO+42, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_DOWN_RIGHT_PEN = to_pen{tile=SBSO+43, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_2TALL_UP_LEFT_PEN = to_pen{tile=SBSO+26, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_2TALL_UP_RIGHT_PEN = to_pen{tile=SBSO+27, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_2TALL_DOWN_LEFT_PEN = to_pen{tile=SBSO+38, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_2TALL_DOWN_RIGHT_PEN = to_pen{tile=SBSO+39, ch=219, fg=COLOR_CYAN, bg=COLOR_BLACK}
-local SCROLLBAR_UP_LEFT_HOVER_PEN = to_pen{tile=SBSO+2, ch=47, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_UP_RIGHT_HOVER_PEN = to_pen{tile=SBSO+3, ch=92, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_DOWN_LEFT_HOVER_PEN = to_pen{tile=SBSO+14, ch=92, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_DOWN_RIGHT_HOVER_PEN = to_pen{tile=SBSO+15, ch=47, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_UP_LEFT_HOVER_PEN = to_pen{tile=SBSO+8, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_UP_RIGHT_HOVER_PEN = to_pen{tile=SBSO+9, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_LEFT_HOVER_PEN = to_pen{tile=SBSO+32, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_RIGHT_HOVER_PEN = to_pen{tile=SBSO+33, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_UP_LEFT_HOVER_PEN = to_pen{tile=SBSO+34, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_UP_RIGHT_HOVER_PEN = to_pen{tile=SBSO+35, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_DOWN_LEFT_HOVER_PEN = to_pen{tile=SBSO+46, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_DOWN_RIGHT_HOVER_PEN = to_pen{tile=SBSO+47, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_LEFT_HOVER_PEN = to_pen{tile=SBSO+20, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_CENTER_RIGHT_HOVER_PEN = to_pen{tile=SBSO+21, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_DOWN_LEFT_HOVER_PEN = to_pen{tile=SBSO+44, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_DOWN_RIGHT_HOVER_PEN = to_pen{tile=SBSO+45, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_2TALL_UP_LEFT_HOVER_PEN = to_pen{tile=SBSO+28, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_2TALL_UP_RIGHT_HOVER_PEN = to_pen{tile=SBSO+29, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_2TALL_DOWN_LEFT_HOVER_PEN = to_pen{tile=SBSO+40, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_2TALL_DOWN_RIGHT_HOVER_PEN = to_pen{tile=SBSO+41, ch=219, fg=COLOR_LIGHTCYAN, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_BG_LEFT_PEN = to_pen{tile=SBSO+12, ch=176, fg=COLOR_DARKGREY, bg=COLOR_BLACK}
-local SCROLLBAR_BAR_BG_RIGHT_PEN = to_pen{tile=SBSO+13, ch=176, fg=COLOR_DARKGREY, bg=COLOR_BLACK}
-
-function Scrollbar:onRenderBody(dc)
-    -- don't draw if all elements are visible
-    if not scrollbar_is_visible(self) then
-        return
-    end
-    -- determine which elements should be highlighted
-    local _,hover_y = self:getMousePos()
-    local hover_up, hover_down, hover_bar = false, false, false
-    if hover_y == 0 then
-        hover_up = true
-    elseif hover_y == dc.height-1 then
-        hover_down = true
-    elseif hover_y then
-        hover_bar = true
-    end
-    -- render up arrow
-    dc:seek(0, 0)
-    dc:char(nil, hover_up and SCROLLBAR_UP_LEFT_HOVER_PEN or SCROLLBAR_UP_LEFT_PEN)
-    dc:char(nil, hover_up and SCROLLBAR_UP_RIGHT_HOVER_PEN or SCROLLBAR_UP_RIGHT_PEN)
-    -- render scrollbar body
-    local starty = self.bar_offset + 1
-    local endy = self.bar_offset + self.bar_height
-    local midy = (starty + endy)/2
-    for y=1,dc.height-2 do
-        dc:seek(0, y)
-        if y >= starty and y <= endy then
-            if y == starty and y <= midy - 1 then
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_UP_LEFT_HOVER_PEN or SCROLLBAR_BAR_UP_LEFT_PEN)
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_UP_RIGHT_HOVER_PEN or SCROLLBAR_BAR_UP_RIGHT_PEN)
-            elseif y == midy - 0.5 and self.bar_height == 2 then
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_2TALL_UP_LEFT_HOVER_PEN or SCROLLBAR_BAR_2TALL_UP_LEFT_PEN)
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_2TALL_UP_RIGHT_HOVER_PEN or SCROLLBAR_BAR_2TALL_UP_RIGHT_PEN)
-            elseif y == midy + 0.5 and self.bar_height == 2 then
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_2TALL_DOWN_LEFT_HOVER_PEN or SCROLLBAR_BAR_2TALL_DOWN_LEFT_PEN)
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_2TALL_DOWN_RIGHT_HOVER_PEN or SCROLLBAR_BAR_2TALL_DOWN_RIGHT_PEN)
-            elseif y == midy - 0.5 then
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_CENTER_UP_LEFT_HOVER_PEN or SCROLLBAR_BAR_CENTER_UP_LEFT_PEN)
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_CENTER_UP_RIGHT_HOVER_PEN or SCROLLBAR_BAR_CENTER_UP_RIGHT_PEN)
-            elseif y == midy + 0.5 then
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_CENTER_DOWN_LEFT_HOVER_PEN or SCROLLBAR_BAR_CENTER_DOWN_LEFT_PEN)
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_CENTER_DOWN_RIGHT_HOVER_PEN or SCROLLBAR_BAR_CENTER_DOWN_RIGHT_PEN)
-            elseif y == midy then
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_CENTER_LEFT_HOVER_PEN or SCROLLBAR_BAR_CENTER_LEFT_PEN)
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_CENTER_RIGHT_HOVER_PEN or SCROLLBAR_BAR_CENTER_RIGHT_PEN)
-            elseif y == endy and y >= midy + 1 then
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_DOWN_LEFT_HOVER_PEN or SCROLLBAR_BAR_DOWN_LEFT_PEN)
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_DOWN_RIGHT_HOVER_PEN or SCROLLBAR_BAR_DOWN_RIGHT_PEN)
-            else
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_LEFT_HOVER_PEN or SCROLLBAR_BAR_LEFT_PEN)
-                dc:char(nil, hover_bar and SCROLLBAR_BAR_RIGHT_HOVER_PEN or SCROLLBAR_BAR_RIGHT_PEN)
-            end
-        else
-            dc:char(nil, SCROLLBAR_BAR_BG_LEFT_PEN)
-            dc:char(nil, SCROLLBAR_BAR_BG_RIGHT_PEN)
-        end
-    end
-    -- render down arrow
-    dc:seek(0, dc.height-1)
-    dc:char(nil, hover_down and SCROLLBAR_DOWN_LEFT_HOVER_PEN or SCROLLBAR_DOWN_LEFT_PEN)
-    dc:char(nil, hover_down and SCROLLBAR_DOWN_RIGHT_HOVER_PEN or SCROLLBAR_DOWN_RIGHT_PEN)
-    if not self.on_scroll then return end
-    -- manage state for dragging and continuous scrolling
-    if self.is_dragging then
-        scrollbar_do_drag(self)
-    end
-    if df.global.enabler.mouse_lbut_down == 0 then
-        self.last_scroll_ms = 0
-        self.is_dragging = false
-        self.scroll_spec = nil
-        return
-    end
-    if self.last_scroll_ms == 0 then return end
-    local now = dfhack.getTickCount()
-    local delay = self.is_first_click and
-            SCROLL_INITIAL_DELAY_MS or SCROLL_DELAY_MS
-    if now - self.last_scroll_ms >= delay then
-        self.is_first_click = false
-        self.on_scroll(self.scroll_spec)
-        self.last_scroll_ms = now
-    end
-end
-
-function Scrollbar:onInput(keys)
-    if not self.on_scroll or not scrollbar_is_visible(self) then
-        return false
-    end
-
-    if self.parent_view and self.parent_view:getMousePos() then
-        if keys.CONTEXT_SCROLL_UP then
-            self.on_scroll('up_small')
-            return true
-        elseif keys.CONTEXT_SCROLL_DOWN then
-            self.on_scroll('down_small')
-            return true
-        elseif keys.CONTEXT_SCROLL_PAGEUP then
-            self.on_scroll('up_large')
-            return true
-        elseif keys.CONTEXT_SCROLL_PAGEDOWN then
-            self.on_scroll('down_large')
-            return true
-        end
-    end
-    if not keys._MOUSE_L then return false end
-    local _,y = self:getMousePos()
-    if not y then return false end
-    local scroll_spec = nil
-    if y == 0 then scroll_spec = 'up_small'
-    elseif y == self.frame_body.height - 1 then scroll_spec = 'down_small'
-    elseif y <= self.bar_offset then scroll_spec = 'up_large'
-    elseif y > self.bar_offset + self.bar_height then scroll_spec = 'down_large'
-    else
-        self.is_dragging = y - self.bar_offset
-        return true
-    end
-    self.scroll_spec = scroll_spec
-    self.on_scroll(scroll_spec)
-    -- reset continuous scroll state
-    self.is_first_click = true
-    self.last_scroll_ms = dfhack.getTickCount()
-    return true
-end
-
------------
--- Label --
------------
-
-function parse_label_text(obj)
-    local text = obj.text or {}
-    if type(text) ~= 'table' then
-        text = { text }
-    end
-    local curline = nil
-    local out = { }
-    local active = nil
-    local idtab = nil
-    for _,v in ipairs(text) do
-        local vv
-        if type(v) == 'string' then
-            vv = v:split(NEWLINE)
-        else
-            vv = { v }
-        end
-
-        for i = 1,#vv do
-            local cv = vv[i]
-            if i > 1 then
-                if not curline then
-                    table.insert(out, {})
-                end
-                curline = nil
-            end
-            if cv ~= '' then
-                if not curline then
-                    curline = {}
-                    table.insert(out, curline)
-                end
-
-                if type(cv) == 'string' then
-                    table.insert(curline, { text = cv })
-                else
-                    table.insert(curline, cv)
-
-                    if cv.on_activate then
-                        active = active or {}
-                        table.insert(active, cv)
-                    end
-
-                    if cv.id then
-                        idtab = idtab or {}
-                        idtab[cv.id] = cv
-                    end
-                end
-            end
-        end
-    end
-    obj.text_lines = out
-    obj.text_active = active
-    obj.text_ids = idtab
-end
-
-local function is_disabled(token)
-    return (token.disabled ~= nil and getval(token.disabled)) or
-           (token.enabled ~= nil and not getval(token.enabled))
-end
-
--- Make the hover pen -- that is a pen that should render elements that has the
--- mouse hovering over it. if hpen is specified, it just checks the fields and
--- returns it (in parsed pen form)
-local function make_hpen(pen, hpen)
-    if not hpen then
-        pen = to_pen(pen)
-
-        -- Swap the foreground and background
-        hpen = dfhack.pen.make(pen.bg, nil, pen.fg + (pen.bold and 8 or 0))
-    end
-
-    -- text_hpen needs a character in order to paint the background using
-    -- Painter:fill(), so let's make it paint a space to show the background
-    -- color
-    local hpen_parsed = to_pen(hpen)
-    hpen_parsed.ch = string.byte(' ')
-    return hpen_parsed
-end
-
----@param obj any
----@param dc gui.Painter
----@param x0 integer
----@param y0 integer
----@param pen dfhack.pen|dfhack.color|fun(): dfhack.pen|dfhack.color
----@param dpen dfhack.pen|dfhack.color|fun(): dfhack.pen|dfhack.color
----@param disabled boolean
----@param hpen dfhack.pen|dfhack.color|fun(): dfhack.pen|dfhack.color
----@param hovered boolean
-function render_text(obj,dc,x0,y0,pen,dpen,disabled,hpen,hovered)
-    pen, dpen, hpen = getval(pen), getval(dpen), getval(hpen)
-    local width = 0
-    for iline = dc and obj.start_line_num or 1, #obj.text_lines do
-        local x, line = 0, obj.text_lines[iline]
-        if dc then
-            local offset = (obj.start_line_num or 1) - 1
-            local y = y0 + iline - offset - 1
-            -- skip text outside of the containing frame
-            if y > dc.height - 1 then break end
-            dc:seek(x+x0, y)
-        end
-        for _,token in ipairs(line) do
-            token.line = iline
-            token.x1 = x
-
-            if token.gap then
-                x = x + token.gap
-                if dc then
-                    dc:advance(token.gap)
-                end
-            end
-
-            if token.tile or (hovered and token.htile) then
-                x = x + 1
-                if dc then
-                    local tile = hovered and getval(token.htile or token.tile) or getval(token.tile)
-                    local tile_pen = tonumber(tile) and to_pen{tile=tile} or tile
-                    dc:char(nil, tile_pen)
-                    if token.width then
-                        dc:advance(token.width-1)
-                    end
-                end
-            end
-
-            if token.text or token.key then
-                local text = ''..(getval(token.text) or '')
-                local keypen = to_pen(token.key_pen or COLOR_LIGHTGREEN)
-
-                if dc then
-                    local tpen = getval(token.pen)
-                    local dcpen = to_pen(tpen or pen)
-
-                    -- If disabled, figure out which dpen to use
-                    if disabled or is_disabled(token) then
-                        dcpen = to_pen(getval(token.dpen) or tpen or dpen)
-                        if keypen.fg ~= COLOR_BLACK then
-                            keypen.bold = false
-                        end
-
-                        -- if hovered *and* disabled, combine both effects
-                        if hovered then
-                            dcpen = make_hpen(dcpen)
-                        end
-                    elseif hovered then
-                        dcpen = make_hpen(dcpen, getval(token.hpen) or hpen)
-                    end
-
-                    dc:pen(dcpen)
-                end
-                local width = getval(token.width)
-                local padstr
-                if width then
-                    x = x + width
-                    if #text > width then
-                        text = string.sub(text,1,width)
-                    else
-                        if token.pad_char then
-                            padstr = string.rep(token.pad_char,width-#text)
-                        end
-                        if dc and token.rjustify then
-                            if padstr then dc:string(padstr) else dc:advance(width-#text) end
-                        end
-                    end
-                else
-                    x = x + #text
-                end
-
-                if token.key then
-                    if type(token.key) == 'string' and not df.interface_key[token.key] then
-                        error('Invalid interface_key: ' .. token.key)
-                    end
-                    local keystr = gui.getKeyDisplay(token.key)
-                    local sep = token.key_sep or ''
-
-                    x = x + #keystr
-
-                    if sep:startswith('()') then
-                        if dc then
-                            dc:string(text)
-                            dc:string(' ('):string(keystr,keypen)
-                            dc:string(sep:sub(2))
-                        end
-                        x = x + 1 + #sep
-                    else
-                        if dc then
-                            dc:string(keystr,keypen):string(sep):string(text)
-                        end
-                        x = x + #sep
-                    end
-                else
-                    if dc then
-                        dc:string(text)
-                    end
-                end
-
-                if width and dc and not token.rjustify then
-                    if padstr then dc:string(padstr) else dc:advance(width-#text) end
-                end
-            end
-
-            token.x2 = x
-        end
-        width = math.max(width, x)
-    end
-    obj.text_width = width
-end
-
-function check_text_keys(self, keys)
-    if self.text_active then
-        for _,item in ipairs(self.text_active) do
-            if item.key and keys[item.key] and not is_disabled(item) then
-                item.on_activate()
-                return true
-            end
-        end
-    end
-end
-
----@class widgets.LabelToken
----@field text string|fun(): string
----@field gap? integer
----@field tile? integer|dfhack.pen
----@field htile? integer|dfhack.pen
----@field width? integer|fun(): integer
----@field pad_char? string
----@field key? string
----@field key_sep? string
----@field disabled? boolean|fun(): boolean
----@field enabled? boolean|fun(): boolean
----@field pen? dfhack.color|dfhack.pen
----@field dpen? dfhack.color|dfhack.pen
----@field hpen? dfhack.color|dfhack.pen
----@field on_activiate? fun()
----@field id? string
----@field line? any Internal use only
----@field x1? any Internal use only
----@field x2? any Internal use only
-
----@class widgets.Label.attrs: widgets.Widget.attrs
----@field text? string|widgets.LabelToken[]
----@field text_pen dfhack.color|dfhack.pen
----@field text_dpen dfhack.color|dfhack.pen
----@field text_hpen? dfhack.color|dfhack.pen
----@field disabled? boolean|fun(): boolean
----@field enabled? boolean|fun(): boolean
----@field auto_height boolean
----@field auto_width boolean
----@field on_click? function
----@field on_rclick? function
----@field scroll_keys table<string, string|integer>
-
----@class widgets.Label.attrs.partial: widgets.Label.attrs
-
----@class widgets.Label: widgets.Widget, widgets.Label.attrs
----@field super widgets.Widget
----@field ATTRS widgets.Label.attrs|fun(attributes: widgets.Label.attrs.partial)
----@overload fun(init_table: widgets.Label.attrs.partial): self
-Label = defclass(Label, Widget)
-
-Label.ATTRS{
-    text_pen = COLOR_WHITE,
-    text_dpen = COLOR_DARKGREY, -- disabled
-    text_hpen = DEFAULT_NIL, -- hover - default is to invert the fg/bg colors
-    disabled = DEFAULT_NIL,
-    enabled = DEFAULT_NIL,
-    auto_height = true,
-    auto_width = false,
-    on_click = DEFAULT_NIL,
-    on_rclick = DEFAULT_NIL,
-    scroll_keys = STANDARDSCROLL,
-}
-
----@param args widgets.Label.attrs.partial
-function Label:init(args)
-    self.scrollbar = Scrollbar{
-        frame={r=0},
-        on_scroll=self:callback('on_scrollbar')}
-
-    self:addviews{self.scrollbar}
-
-    self:setText(args.text or self.text)
-end
-
-local function update_label_scrollbar(label)
-    local body_height = label.frame_body and label.frame_body.height or 1
-    label.scrollbar:update(label.start_line_num, body_height,
-                           label:getTextHeight())
-end
-
-function Label:setText(text)
-    self.start_line_num = 1
-    self.text = text
-    parse_label_text(self)
-
-    if self.auto_height then
-        self.frame = self.frame or {}
-        self.frame.h = self:getTextHeight()
-    end
-
-    update_label_scrollbar(self)
-end
-
-function Label:preUpdateLayout()
-    if self.auto_width then
-        self.frame = self.frame or {}
-        self.frame.w = self:getTextWidth()
-    end
-end
-
-function Label:postUpdateLayout()
-    update_label_scrollbar(self)
-end
-
-function Label:itemById(id)
-    if self.text_ids then
-        return self.text_ids[id]
-    end
-end
-
-function Label:getTextHeight()
-    return #self.text_lines
-end
-
-function Label:getTextWidth()
-    render_text(self)
-    return self.text_width
-end
-
--- Overridden by subclasses that also want to add new mouse handlers, see
--- HotkeyLabel.
-function Label:shouldHover()
-    return self.on_click or self.on_rclick
-end
-
-function Label:onRenderBody(dc)
-    local text_pen = self.text_pen
-    local hovered = self:getMousePos() and self:shouldHover()
-    render_text(self,dc,0,0,text_pen,self.text_dpen,is_disabled(self), self.text_hpen, hovered)
-end
-
-function Label:on_scrollbar(scroll_spec)
-    local v = 0
-    if tonumber(scroll_spec) then
-        v = scroll_spec - self.start_line_num
-    elseif scroll_spec == 'down_large' then
-        v = '+halfpage'
-    elseif scroll_spec == 'up_large' then
-        v = '-halfpage'
-    elseif scroll_spec == 'down_small' then
-        v = 1
-    elseif scroll_spec == 'up_small' then
-        v = -1
-    end
-
-    self:scroll(v)
-end
-
-function Label:scroll(nlines)
-    if not nlines then return end
-    local text_height = math.max(1, self:getTextHeight())
-    if type(nlines) == 'string' then
-        if nlines == '+page' then
-            nlines = self.frame_body.height
-        elseif nlines == '-page' then
-            nlines = -self.frame_body.height
-        elseif nlines == '+halfpage' then
-            nlines = math.ceil(self.frame_body.height/2)
-        elseif nlines == '-halfpage' then
-            nlines = -math.ceil(self.frame_body.height/2)
-        elseif nlines == 'home' then
-            nlines = 1 - self.start_line_num
-        elseif nlines == 'end' then
-            nlines = text_height
-        else
-            error(('unhandled scroll keyword: "%s"'):format(nlines))
-        end
-    end
-    local n = self.start_line_num + nlines
-    n = math.min(n, text_height - self.frame_body.height + 1)
-    n = math.max(n, 1)
-    nlines = n - self.start_line_num
-    self.start_line_num = n
-    update_label_scrollbar(self)
-    return nlines
-end
-
-function Label:onInput(keys)
-    if is_disabled(self) then return false end
-    if self:inputToSubviews(keys) then
-        return true
-    end
-    if keys._MOUSE_L and self:getMousePos() and self.on_click then
-        self.on_click()
-        return true
-    end
-    if keys._MOUSE_R and self:getMousePos() and self.on_rclick then
-        self.on_rclick()
-        return true
-    end
-    for k,v in pairs(self.scroll_keys) do
-        if keys[k] and 0 ~= self:scroll(v) then
-            return true
-        end
-    end
-    return check_text_keys(self, keys)
-end
-
---------------------------------
--- makeButtonLabelText
---
-
-local function get_button_token_hover_ch(spec, x, y, ch)
-    local ch_hover = ch
-    if spec.chars_hover then
-        local row = spec.chars_hover[y]
-        if type(row) == 'string' then
-            ch_hover = row:sub(x, x)
-        else
-            ch_hover = row[x]
-        end
-    end
-    return ch_hover
-end
-
-local function get_button_token_base_pens(spec, x, y)
-    local pen, pen_hover = COLOR_GRAY, COLOR_WHITE
-    if spec.pens then
-        pen = type(spec.pens) == 'table' and (safe_index(spec.pens, y, x) or spec.pens[y]) or spec.pens
-        if spec.pens_hover then
-            pen_hover = type(spec.pens_hover) == 'table' and (safe_index(spec.pens_hover, y, x) or spec.pens_hover[y]) or spec.pens_hover
-        else
-            pen_hover = pen
-        end
-    end
-    return pen, pen_hover
-end
-
-local function get_button_tileset_idx(spec, x, y, tileset_offset, tileset_stride)
-    local idx = (tileset_offset or 1)
-    idx = idx + (x - 1)
-    idx = idx + (y - 1) * (tileset_stride or #spec.chars[1])
-    return idx
-end
-
-local function get_asset_tile(asset, x, y)
-    return dfhack.screen.findGraphicsTile(asset.page, asset.x+x-1, asset.y+y-1)
-end
-
-local function get_button_token_tiles(spec, x, y)
-    local tile = safe_index(spec.tiles_override, y, x)
-    local tile_hover = safe_index(spec.tiles_hover_override, y, x) or tile
-    if not tile and spec.tileset then
-        local tileset = spec.tileset
-        local idx = get_button_tileset_idx(spec, x, y, spec.tileset_offset, spec.tileset_stride)
-        tile = dfhack.textures.getTexposByHandle(tileset[idx])
-        if spec.tileset_hover then
-            local tileset_hover = spec.tileset_hover
-            local idx_hover = get_button_tileset_idx(spec, x, y,
-                spec.tileset_hover_offset or spec.tileset_offset,
-                spec.tileset_hover_stride or spec.tileset_stride)
-            tile_hover = dfhack.textures.getTexposByHandle(tileset_hover[idx_hover])
-        else
-            tile_hover = tile
-        end
-    end
-    if not tile and spec.asset then
-        tile = get_asset_tile(spec.asset, x, y)
-        if spec.asset_hover then
-            tile_hover = get_asset_tile(spec.asset_hover, x, y)
-        else
-            tile_hover = tile
-        end
-    end
-    return tile, tile_hover
-end
-
-local function get_button_token_pen(base_pen, tile, ch)
-    local pen = dfhack.pen.make(base_pen)
-    pen.tile = tile
-    pen.ch = ch
-    return pen
-end
-
-local function get_button_token_pens(spec, x, y, ch, ch_hover)
-    local base_pen, base_pen_hover = get_button_token_base_pens(spec, x, y)
-    local tile, tile_hover = get_button_token_tiles(spec, x, y)
-    return get_button_token_pen(base_pen, tile, ch), get_button_token_pen(base_pen_hover, tile_hover, ch_hover)
-end
-
-local function make_button_token(spec, x, y, ch)
-    local ch_hover = get_button_token_hover_ch(spec, x, y, ch)
-    local pen, pen_hover = get_button_token_pens(spec, x, y, ch, ch_hover)
-    return {
-        tile=pen,
-        htile=pen_hover,
-        width=1,
-    }
-end
-
----@class widgets.ButtonLabelSpec
----@field chars (string|string[])[]
----@field chars_hover? (string|string[])[]
----@field pens? dfhack.color|dfhack.color[][]
----@field pens_hover? dfhack.color|dfhack.color[][]
----@field tiles_override? integer[][]
----@field tiles_hover_override? integer[][]
----@field tileset? TexposHandle[]
----@field tileset_hover? TexposHandle[]
----@field tileset_offset? integer
----@field tileset_hover_offset? integer
----@field tileset_stride? integer
----@field tileset_hover_stride? integer
----@field asset? {page: string, x: integer, y: integer}
----@field asset_hover? {page: string, x: integer, y: integer}
-
----@nodiscard
----@param spec widgets.ButtonLabelSpec
----@return widgets.LabelToken[]
-function makeButtonLabelText(spec)
-    local tokens = {}
-    for y, row in ipairs(spec.chars) do
-        if type(row) == 'string' then
-            local x = 1
-            for ch in row:gmatch('.') do
-                table.insert(tokens, make_button_token(spec, x, y, ch))
-                x = x + 1
-            end
-        else
-            for x, ch in ipairs(row) do
-                table.insert(tokens, make_button_token(spec, x, y, ch))
-            end
-        end
-        if y < #spec.chars then
-            table.insert(tokens, NEWLINE)
-        end
-    end
-    return tokens
 end
 
 ------------------
@@ -1233,65 +108,6 @@ TooltipLabel.ATTRS{
 
 function TooltipLabel:init()
     self.visible = self.show_tooltip
-end
-
------------------
--- HotkeyLabel --
------------------
-
----@class widgets.HotkeyLabel.attrs: widgets.Label.attrs
----@field key? string
----@field key_sep string
----@field label? string|fun(): string
----@field on_activate? function
-
----@class widgets.HotkeyLabel.attrs.partial: widgets.HotkeyLabel.attrs
-
----@class widgets.HotkeyLabel: widgets.Label, widgets.HotkeyLabel.attrs
----@field super widgets.Label
----@field ATTRS widgets.HotkeyLabel.attrs|fun(attributes: widgets.HotkeyLabel.attrs.partial)
----@overload fun(init_table: widgets.HotkeyLabel.attrs.partial): self
-HotkeyLabel = defclass(HotkeyLabel, Label)
-
-HotkeyLabel.ATTRS{
-    key=DEFAULT_NIL,
-    key_sep=': ',
-    label=DEFAULT_NIL,
-    on_activate=DEFAULT_NIL,
-}
-
-function HotkeyLabel:init()
-    self:initializeLabel()
-end
-
-function HotkeyLabel:setOnActivate(on_activate)
-    self.on_activate = on_activate
-    self:initializeLabel()
-end
-
-function HotkeyLabel:setLabel(label)
-    self.label = label
-    self:initializeLabel()
-end
-
-function HotkeyLabel:shouldHover()
-    -- When on_activate is set, text should also hover on mouseover
-    return self.on_activate or HotkeyLabel.super.shouldHover(self)
-end
-
-function HotkeyLabel:initializeLabel()
-    self:setText{{key=self.key, key_sep=self.key_sep, text=self.label,
-                  on_activate=self.on_activate}}
-end
-
-function HotkeyLabel:onInput(keys)
-    if HotkeyLabel.super.onInput(self, keys) then
-        return true
-    elseif keys._MOUSE_L and self:getMousePos() and self.on_activate
-            and not is_disabled(self) then
-        self.on_activate()
-        return true
-    end
 end
 
 ----------------
